% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highs.R
\name{highs_optimizer}
\alias{highs_optimizer}
\title{Solve {ompr} models using HiGHS}
\usage{
highs_optimizer(control = list())
}
\arguments{
\item{control}{a list of options passed to \code{highs::highs_solve}. For
a complete list see \code{highs::highs_available_solver_options()}.}
}
\value{
a function: Model -> Solution that can be used
together with \code{\link[ompr]{solve_model}}.
}
\description{
The solver uses {highs} internally to solve MIP models.
}
\examples{
\dontrun{
library(magrittr)
library(ompr)
library(ROI)
library(ROI.plugin.glpk)
add_variable(MIPModel(), x, type = "continuous") \%>\%
    set_objective(x, sense = "max") \%>\%
    add_constraint(x <= 5) \%>\%
    solve_model(highs_optimizer())
}
}
